--- a/cmd/agent/api/grpc.go
+++ b/cmd/agent/api/grpc.go
@@ -12,17 +12,18 @@ package api
 
 import (
 	"context"
+	"fmt"
+	"strings"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 
 	pb "github.com/DataDog/datadog-agent/cmd/agent/api/pb"
 	"github.com/DataDog/datadog-agent/pkg/tagger"
 	"github.com/DataDog/datadog-agent/pkg/tagger/collectors"
 	hostutil "github.com/DataDog/datadog-agent/pkg/util"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
-
-// EnableExperimentalEndpoints enables experimental endpoints (set via -X)
-var EnableExperimentalEndpoints bool
 
 type server struct {
 	pb.UnimplementedAgentServer
@@ -30,6 +31,11 @@ type server struct {
 
 type serverSecure struct {
 	pb.UnimplementedAgentServer
+
+	// NOTE: tagger.Tagger is a concrete type that makes testing harder
+	// than it should be. We should make that concrete type private, and
+	// create a new tagger.Tagger interface that replicates it.
+	tagger *tagger.Tagger
 }
 
 func (s *server) GetHostname(ctx context.Context, in *pb.HostnameRequest) (*pb.HostnameReply, error) {
@@ -48,12 +54,118 @@ func (s *server) AuthFuncOverride(ctx context.Context, fullMethodName string) (c
 	return ctx, nil
 }
 
-func (s *serverSecure) GetTags(ctx context.Context, in *pb.TagRequest) (*pb.TagReply, error) {
-	if EnableExperimentalEndpoints {
-		tags, _ := tagger.Tag(in.GetEntity(), collectors.HighCardinality)
-		return &pb.TagReply{Tags: tags}, nil
+// StreamTags subscribes to added, removed, or changed entities in the Tagger
+// and streams them to clients as pb.StreamTagsResponse events. Filtering is as
+// of yet not implemented.
+func (s *serverSecure) TaggerStreamEntities(in *pb.StreamTagsRequest, out pb.AgentSecure_TaggerStreamEntitiesServer) error {
+	cardinality, err := pb2taggerCardinality(in.Cardinality)
+	if err != nil {
+		return err
 	}
 
-	return nil, status.Errorf(codes.PermissionDenied,
-		"This is an experimental endpoint and has been disabled in this build")
+	// NOTE: StreamTagsRequest can specify filters, but they cannot be
+	// implemented since the tagger has no concept of container metadata.
+	// these filters will be introduced when we implement a container
+	// metadata service that can receive them as is from the tagger.
+
+	eventCh := s.tagger.Subscribe(cardinality)
+	defer s.tagger.Unsubscribe(eventCh)
+
+	for events := range eventCh {
+		for _, event := range events {
+			response, err := tagger2pbEntityEvent(event)
+			if err != nil {
+				log.Warnf("can't convert tagger entity to protobuf: %s", err)
+				continue
+			}
+
+			err = out.Send(response)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
+
+// FetchEntity fetches an entity from the Tagger with the desired cardinality tags.
+func (s *serverSecure) TaggerFetchEntity(ctx context.Context, in *pb.FetchEntityRequest) (*pb.FetchEntityResponse, error) {
+	if in.Id == nil {
+		return nil, status.Errorf(codes.InvalidArgument, `missing "id" parameter`)
+	}
+
+	entityID := fmt.Sprintf("%s://%s", in.Id.Prefix, in.Id.Uid)
+	cardinality, err := pb2taggerCardinality(in.Cardinality)
+	if err != nil {
+		return nil, err
+	}
+
+	tags, err := s.tagger.Tag(entityID, cardinality)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "%s", err)
+	}
+
+	return &pb.FetchEntityResponse{
+		Id:          in.Id,
+		Cardinality: in.Cardinality,
+		Tags:        tags,
+	}, nil
+}
+
+func tagger2pbEntityID(entityID string) (*pb.EntityId, error) {
+	parts := strings.SplitN(entityID, "://", 2)
+	if len(parts) != 2 {
+		return nil, fmt.Errorf("invalid entity id %q", entityID)
+	}
+
+	return &pb.EntityId{
+		Prefix: parts[0],
+		Uid:    parts[1],
+	}, nil
+}
+
+func tagger2pbEntityEvent(event tagger.EntityEvent) (*pb.StreamTagsResponse, error) {
+	entity := event.Entity
+	entityID, err := tagger2pbEntityID(entity.ID)
+	if err != nil {
+		return nil, err
+	}
+
+	var eventType pb.EventType
+	switch event.EventType {
+	case tagger.EventTypeAdded:
+		eventType = pb.EventType_ADDED
+	case tagger.EventTypeModified:
+		eventType = pb.EventType_MODIFIED
+	case tagger.EventTypeDeleted:
+		eventType = pb.EventType_DELETED
+	default:
+		return nil, fmt.Errorf("invalid event type %q", event.EventType)
+	}
+
+	return &pb.StreamTagsResponse{
+		Type: eventType,
+		Entity: &pb.Entity{
+			Id:                          entityID,
+			Hash:                        entity.Hash,
+			HighCardinalityTags:         entity.HighCardinalityTags,
+			OrchestratorCardinalityTags: entity.OrchestratorCardinalityTags,
+			LowCardinalityTags:          entity.LowCardinalityTags,
+			StandardTags:                entity.StandardTags,
+		},
+	}, nil
+}
+
+func pb2taggerCardinality(pbCardinality pb.TagCardinality) (collectors.TagCardinality, error) {
+	switch pbCardinality {
+	case pb.TagCardinality_LOW:
+		return collectors.LowCardinality, nil
+	case pb.TagCardinality_ORCHESTRATOR:
+		return collectors.OrchestratorCardinality, nil
+	case pb.TagCardinality_HIGH:
+		return collectors.HighCardinality, nil
+	}
+
+	return 0, status.Errorf(codes.InvalidArgument, "invalid cardinality %q", pbCardinality)
 }