--- a/cmd/agent/api/pb/api.proto
+++ b/cmd/agent/api/pb/api.proto
@@ -15,12 +15,57 @@ service Agent {
 }
 
 service AgentSecure {
-    rpc GetTags (TagRequest) returns (TagReply) {
+    // subscribes to added, removed, or changed entities in the Tagger
+    // and streams them to clients as events.
+    // can be called through the HTTP gateway, and events will be streamed as JSON:
+    //   $  curl -H "authorization: Bearer $(cat /etc/datadog-agent/auth_token)" \
+    //      -XPOST -k https://localhost:5001/v1/grpc/tagger/stream_entities
+    //   {
+    //    "result": {
+    //        "entity": {
+    //            "id": {
+    //                "prefix": "kubernetes_pod_uid",
+    //                "uid": "4025461f832caf3fceb7fc2a32f879c6"
+    //            },
+    //            "hash": "cad4fc8fc409fcc1",
+    //            "lowCardinalityTags": [
+    //                "kube_namespace:kube-system",
+    //                "pod_phase:running"
+    //            ]
+    //        }
+    //    }
+    //}
+    rpc TaggerStreamEntities(StreamTagsRequest) returns (stream StreamTagsResponse) {
         option (google.api.http) = {
-            post: "/v1/grpc/tags"
+            post: "/v1/grpc/tagger/stream_entities"
             body: "*"
         };
-    }
+    };
+
+    // fetches an entity from the Tagger with the desired cardinality tags.
+    // can be called through the HTTP gateway, and entity will be returned as JSON:
+    //   $ curl -H "authorization: Bearer $(cat /etc/datadog-agent/auth_token)" \
+    //      -XPOST -k -H "Content-Type: application/json" \
+    //      --data '{"id":{"prefix":"kubernetes_pod_uid","uid":"d575fb58-82dc-418e-bfb1-aececc9bc507"}}' \
+    //      https://localhost:5001/v1/grpc/tagger/fetch_entity 
+    //   {
+    //    "id": {
+    //        "prefix": "kubernetes_pod_uid",
+    //        "uid": "d575fb58-82dc-418e-bfb1-aececc9bc507"
+    //    },
+    //    "tags": [
+    //        "kube_namespace:kube-system",
+    //        "pod_phase:running",
+    //        "kube_deployment:coredns",
+    //        "kube_service:kube-dns"
+    //    ]
+    //}
+    rpc TaggerFetchEntity(FetchEntityRequest) returns (FetchEntityResponse) {
+        option (google.api.http) = {
+            post: "/v1/grpc/tagger/fetch_entity"
+            body: "*"
+        };
+    };
 }
 
 message HostnameRequest {}
@@ -30,12 +75,56 @@ message HostnameReply {
     string hostname = 1;
 }
 
-// The request message containing the tag list for an entity.
-message TagRequest {
-    string entity = 1;
+message StreamTagsRequest {
+    TagCardinality cardinality = 1;
+    Filter includeFilter = 2;
+    Filter excludeFilter = 3;
 }
 
-// The response message containing the tagger reply
-message TagReply {
-    repeated string tags = 1;
+message StreamTagsResponse {
+    EventType type = 1;
+    Entity entity = 2;
+}
+
+enum EventType {
+    ADDED = 0;
+    MODIFIED = 1;
+    DELETED = 2;
+}
+
+enum TagCardinality {
+    LOW = 0;
+    ORCHESTRATOR = 1;
+    HIGH = 2;
+}
+
+message Filter {
+    string kubeNamespace = 1;
+    string image = 2;
+    string containerName = 3;
+}
+
+message Entity {
+    EntityId id = 1;
+    string hash = 2;
+    repeated string highCardinalityTags = 3;
+    repeated string orchestratorCardinalityTags = 4;
+    repeated string lowCardinalityTags = 5;
+    repeated string standardTags = 6;
+}
+
+message FetchEntityRequest {
+    EntityId id = 1;
+    TagCardinality cardinality = 2;
+}
+
+message FetchEntityResponse {
+    EntityId id = 1;
+    TagCardinality cardinality = 2;
+    repeated string tags = 3;
+}
+
+message EntityId {
+    string prefix = 1;
+    string uid = 2;
 }