--- a/pkg/tagger/tagger_test.go
+++ b/pkg/tagger/tagger_test.go
@@ -127,16 +127,18 @@ func TestFetchAllCached(t *testing.T) {
 	tagger.Init(catalog)
 	defer tagger.Stop()
 
-	tagger.tagStore.processTagInfo(&collectors.TagInfo{
-		Entity:       "entity_name",
-		Source:       "stream",
-		LowCardTags:  []string{"low1"},
-		HighCardTags: []string{"high"},
-	})
-	tagger.tagStore.processTagInfo(&collectors.TagInfo{
-		Entity:      "entity_name",
-		Source:      "pull",
-		LowCardTags: []string{"low2"},
+	tagger.tagStore.processTagInfo([]*collectors.TagInfo{
+		{
+			Entity:       "entity_name",
+			Source:       "stream",
+			LowCardTags:  []string{"low1"},
+			HighCardTags: []string{"high"},
+		},
+		{
+			Entity:      "entity_name",
+			Source:      "pull",
+			LowCardTags: []string{"low2"},
+		},
 	})
 
 	streamer := tagger.streamers["stream"].(*DummyCollector)
@@ -172,10 +174,12 @@ func TestFetchOneCached(t *testing.T) {
 	tagger.Init(catalog)
 	defer tagger.Stop()
 
-	tagger.tagStore.processTagInfo(&collectors.TagInfo{
-		Entity:      "entity_name",
-		Source:      "stream",
-		LowCardTags: []string{"low1"},
+	tagger.tagStore.processTagInfo([]*collectors.TagInfo{
+		{
+			Entity:      "entity_name",
+			Source:      "stream",
+			LowCardTags: []string{"low1"},
+		},
 	})
 
 	streamer := tagger.streamers["stream"].(*DummyCollector)
@@ -207,10 +211,12 @@ func TestEmptyEntity(t *testing.T) {
 	tagger.Init(catalog)
 	defer tagger.Stop()
 
-	tagger.tagStore.processTagInfo(&collectors.TagInfo{
-		Entity:      "entity_name",
-		Source:      "stream",
-		LowCardTags: []string{"low1"},
+	tagger.tagStore.processTagInfo([]*collectors.TagInfo{
+		{
+			Entity:      "entity_name",
+			Source:      "stream",
+			LowCardTags: []string{"low1"},
+		},
 	})
 
 	tags, err := tagger.Tag("", collectors.HighCardinality)
@@ -297,10 +303,12 @@ func TestSafeCache(t *testing.T) {
 	tagger.Init(catalog)
 	defer tagger.Stop()
 
-	tagger.tagStore.processTagInfo(&collectors.TagInfo{
-		Entity:      "entity_name",
-		Source:      "pull",
-		LowCardTags: []string{"low1", "low2", "low3"},
+	tagger.tagStore.processTagInfo([]*collectors.TagInfo{
+		{
+			Entity:      "entity_name",
+			Source:      "pull",
+			LowCardTags: []string{"low1", "low2", "low3"},
+		},
 	})
 
 	// First lookup