… cmd/agent/api/pb/api.pb.go
… pkg/tagger/tagstore.go
… pkg/tagger/tagstore_test.go