### What does this PR do?

This PR introduces a gRPC interface that wraps the Tagger and exposes two RPCs: `StreamTags` to watch and stream changes to entities in a "push" model, and `FetchEntity` to pull the current tags for one entity.

### Motivation

This is the first part in an epic that aims to centralize the Tagger into a single agent, instead of having one per agent.

### Describe your test plan

The changes here can probably be tested more easily when we introduce the other side of this equation, by implementing the client for this. If you still want to give it a go, I think [grpcurl](https://github.com/fullstorydev/grpcurl) can work.