@juliogreff 2020-10-20
I added some benchmarks to ensure we don't let performance degrade too much. Below are the results. Benchmarks were run with `go test -run=NONE -bench=TagStore -count 5 ./pkg/tagger`.

```shell
$ benchstat old.txt new.txt
name                       old time/op  new time/op  delta
TagStoreThroughput-8        307ms ± 2%   322ms ± 2%  +4.70%  (p=0.008 n=5+5)
TagStore_processTagInfo-8   123µs ±12%   126µs ±12%    ~     (p=0.690 n=5+5)
```

`processTagInfo` being run on its own basically shows no change. The overall throughput is a little bit worse (4.7%), but that's expected since there's extra coordination required to notify subscribers.